# backend/openapi/admin.openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/multipart/{bucket}/initiate": {
      "post": {
        "operationId": "MultipartController_initiate",
        "parameters": [
          {
            "name": "bucket",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InitiateMultipartDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "multipart"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/multipart/{bucket}/{uploadId}/part/{partNumber}": {
      "put": {
        "operationId": "MultipartController_uploadPart",
        "parameters": [
          {
            "name": "bucket",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "uploadId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "partNumber",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": [
                  "part"
                ],
                "properties": {
                  "part": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "multipart"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/multipart/{bucket}/{uploadId}/parts": {
      "get": {
        "operationId": "MultipartController_listParts",
        "parameters": [
          {
            "name": "bucket",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "uploadId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListPartsResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "multipart"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/multipart/{bucket}/{uploadId}/complete": {
      "post": {
        "operationId": "MultipartController_complete",
        "parameters": [
          {
            "name": "bucket",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "uploadId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompleteMultipartDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "multipart"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/multipart/{bucket}/{uploadId}": {
      "delete": {
        "operationId": "MultipartController_abort",
        "parameters": [
          {
            "name": "bucket",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "uploadId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "ok": {
                      "type": "boolean"
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [
          "multipart"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/multipart/{bucket}/uploads": {
      "get": {
        "operationId": "MultipartController_listUploads",
        "parameters": [
          {
            "name": "bucket",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "multipart"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/lifecycle/{bucket}": {
      "get": {
        "operationId": "LifecycleController_get",
        "parameters": [
          {
            "name": "bucket",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetBucketLifecycleResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "lifecycle"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "put": {
        "operationId": "LifecycleController_put",
        "parameters": [
          {
            "name": "bucket",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PutBucketLifecycleDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetBucketLifecycleResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "lifecycle"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "operationId": "LifecycleController_remove",
        "parameters": [
          {
            "name": "bucket",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "ok": {
                      "type": "boolean"
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [
          "lifecycle"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    }
  },
  "info": {
    "title": "LV S3 Admin API",
    "description": "Admin console API for bucket and identity management",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "InitiateMultipartDto": {
        "type": "object",
        "properties": {
          "key": {
            "type": "string",
            "description": "Target object key"
          },
          "partSize": {
            "type": "number",
            "description": "Preferred part size in bytes",
            "minimum": 5242880
          },
          "totalSize": {
            "type": "number",
            "description": "File size in bytes"
          }
        },
        "required": [
          "key"
        ]
      },
      "MultipartPartItemDto": {
        "type": "object",
        "properties": {
          "partNumber": {
            "type": "number"
          },
          "size": {
            "type": "number"
          },
          "etag": {
            "type": "string"
          }
        },
        "required": [
          "partNumber",
          "size",
          "etag"
        ]
      },
      "ListPartsResponseDto": {
        "type": "object",
        "properties": {
          "uploadId": {
            "type": "string"
          },
          "bucket": {
            "type": "string"
          },
          "key": {
            "type": "string"
          },
          "parts": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MultipartPartItemDto"
            }
          }
        },
        "required": [
          "uploadId",
          "bucket",
          "key",
          "parts"
        ]
      },
      "CompleteMultipartDto": {
        "type": "object",
        "properties": {
          "partNumbers": {
            "description": "Ordered part numbers to finalize",
            "type": "array",
            "items": {
              "type": "number"
            }
          },
          "key": {
            "type": "string",
            "description": "Final key (must match initiated key)"
          }
        },
        "required": [
          "partNumbers",
          "key"
        ]
      },
      "LifecycleRuleDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Client-defined stable rule id"
          },
          "enabled": {
            "type": "boolean",
            "description": "Whether this rule is active"
          },
          "prefix": {
            "type": "string",
            "description": "Applies only to object keys starting with this prefix"
          },
          "expirationDays": {
            "type": "number",
            "description": "Delete objects older than N days",
            "minimum": 1,
            "maximum": 3650
          },
          "abortMultipartAfterDays": {
            "type": "number",
            "description": "Abort multipart uploads older than N days",
            "minimum": 1,
            "maximum": 3650
          }
        },
        "required": [
          "id",
          "enabled"
        ]
      },
      "GetBucketLifecycleResponseDto": {
        "type": "object",
        "properties": {
          "bucket": {
            "type": "string"
          },
          "rules": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LifecycleRuleDto"
            }
          }
        },
        "required": [
          "bucket",
          "rules"
        ]
      },
      "PutBucketLifecycleDto": {
        "type": "object",
        "properties": {
          "rules": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LifecycleRuleDto"
            }
          }
        },
        "required": [
          "rules"
        ]
      }
    }
  }
}